backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database (Async)
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

# Pydantic & Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Background Tasks
celery==5.4.0
redis==5.1.1

# Data Processing
pandas==2.2.3
numpy==2.1.2

# Templates & Static Files
jinja2==3.1.4
aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0

# Development
httpx==0.27.2  # For testing
pytest==8.3.3
pytest-asyncio==0.24.0